remove changelog-types.d.ts and add them to parsePRDescription.ts
