+++ dangerfile.ts
@@ -1,9 +1,8 @@
 import * as fs from "fs"
 import { danger, fail, markdown, warn } from "danger"
 import { pickBy } from "lodash"
-import { ParseResult } from "./scripts/changelog/changelog-types"
 import { changelogTemplateSections } from "./scripts/changelog/generateChangelogSectionTemplate"
-import { parsePRDescription } from "./scripts/changelog/parsePRDescription"
+import { ParseResult, parsePRDescription } from "./scripts/changelog/parsePRDescription"
 // TypeScript thinks we're in React Native,
 // so the node API gives us errors:
